Change PRNG fault model to "voltage droop"-like
This uses a new fault model for the PRNG target that uses multiple
parameters to determine if and when faults are injected:
  * Probability of a voltage event
  * Duration of a voltage event
  * Probability of a fault during a voltage event

When a voltage event occurs (determined by a PRNG with a programmable
difficulty inside LeChiffre) then faults are enabled for all
instrumented registers for the duration of the voltage event. A new
voltage event will reset the counter for the voltage event ending.

These parameters are communicated via new psuedo-CSR read/write
operations of the LeChiffre ABI.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>