Inject in Chisel, setup replace signals in FIRRTL
This splits the way that the instrumentation is done. The
instrumentation to add fault injection logic is supplied when
components are tagged from the method of a Chisel Module trait. The
info collection for FIRRTL-based signal replacement (substitution of
possibly faulty components for their originals) is then completed
during annotation collection and intended to be handled by a
to-be-completed FIRRTL pass.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>