Use Injector class, initial hook up logic
This adds an abstract class describing a Chiffre Injector module that
is added on a per-bit basis to each component that has fault injection
capabilities. A concrete example is provided of an LFSR injector. The
injector is then added at the Chisel level, but left unconnected.

This adds a (Verilog spot-checked) transformation that:
  * Hooks up the original signal to the injector input
  * Connects the injector faulty output in place of the original
    signal

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>