Expect a FIRRTLException instead
Before applying this PR, Chiffre did not support non-Ground types and
would throw a FaultInstrumentationException if one was found.  As of
this PR, that is no longer the case, but the test checking for the error
uses a 0-width array which does still throw an exception.  Since I
believe this should be expected behavior, I changed the expected
exception to be the one thrown by fromBits when trying to concatenate a
0-width array of signals.