Improve, fix ScanField equality
This improves ScanField equality with a better hashcode dependent on the
value of a ScanField and not the width of a ScanField (this should cause
fewer collisions). The canEqual method is made abstract in ScanField as
this will be better implemented when mixed into a case class. The equality
check now uses canEqual to do better case class object comparison (two
different case classes with the same value and width will not be equal
unless they are of the same type).

This adds tests comparing ScanFields to verify correct behavior of the
following three situations:
  - Two case classes of the same type and width
  - Two case classes of the same type and different widths
  - Two case classes of different types and the same width

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>