Both ScanField and InjectorInfo now extend Equals
The need for an abstrct `canEqual` method is properly met by mixing in the
scala.Equals trait. This was previously done manually with an abstract
method.

Tests are updated to support this where needed. A few uses of a class
extending InjectorInfo were updated to be case objects.

Note: the case objects currently do not serialize (but they do not need to
for the available tests).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>