Fix InjectorInfo.equality, and InjectorInfo.bind
This changes the `equals` method of `InjectorInfo` to use `canEqual`. This
corrects an error with `InjectorInfo` equality where types with the same
structure (the exact same scan field types and values) would be equal.

This additionally adds an `InjectorInfo.bind(x: Seq[ScanField])` method to
facilitate binding a sequence of scan fields in one fell swoop. Towards
providing a clearer exception on binding failure, this adds a
`ScanFieldBindingException` and uses this in both `ScanField` and
`InjectorInfo` binding methods.

This includes tests that `InjectorInfo` binding and equality work. The
latter is modeled off of `ScanField` equality tests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>