Make isFaulty only work for Data
This makes `isFaulty` only work on Data as opposed to arbitrary
InstanceId. The reason for this is twofold: (1) InstanceId is in
chisel3.internal and it's best to avoid its use if possible and (2) if
anything other than Data was annotated as being faulty, this would trow an
exception. It's better to handle this with a static type check.

Additionally, the ChiffreInjectee trait is made to use a self-type as
opposed to inheritance and a self type (not sure why it was doing this?).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>